+++ client/codedrills/src/features/practice/Practice.tsx
@@ -1,69 +1,26 @@
-import { useEffect, useState } from "react";
-import type { Result, ResultPayload } from "../../types/types";
-import { sendResult } from "../../api/data";
-import { useTasks } from "../../context/tasksContext";
+import { useRef } from "react";
+import { usePracticeSession } from "./usePracticeSession";
 
 export default function Practice() {
-  const { tasks, currentBranch, loading, error, setError, getTasks } =
-    useTasks();
+  const {
+    loading,
+    updating,
+    error,
+    branchTasks,
+    currentTask,
+    currentPrompt,
+    level,
+    answer,
+    updateAnswer,
+    submitAnswer,
+    totalLevels,
+  } = usePracticeSession();
 
-  const branchTasks = tasks[currentBranch] ?? [];
-
-  const [level, setLevel] = useState(0);
-  const [current, setCurrent] = useState(0);
-  const [answer, setAnswer] = useState("");
-  const [done, setDone] = useState(false);
-
-  const currentTask = branchTasks[current];
-  const currentPrompt = currentTask?.levels[level] ?? "";
-
-  useEffect(() => {
-    setLevel(0);
-    setCurrent(0);
-    setAnswer("");
-    setDone(false);
-    setError(null);
-  }, [currentBranch, setError]);
-
-  async function onComplete(payload: ResultPayload) {
-    await sendResult(payload);
-    await getTasks();
-  }
+  const editorRef = useRef<HTMLDivElement | null>(null);
 
   async function handleSubmit(event: React.SubmitEvent<HTMLFormElement>) {
     event.preventDefault();
-
-    const isCorrect = currentTask.code.trim() === answer.trim();
-
-    if (!isCorrect) {
-      setError("Try again!");
-      return;
-    }
-
-    setAnswer("");
-    setError(null);
-
-    const isLastLevel = level === 2;
-    if (!isLastLevel) {
-      setLevel((prev) => prev + 1);
-      return;
-    }
-
-    const completedResult = { title: currentTask.title, done: true };
-    const isLastTask = current === branchTasks.length - 1;
-
-    await onComplete({
-      branch: currentBranch,
-      results: [completedResult],
-    });
-
-    if (isLastTask) {
-      setDone(true);
-      return;
-    }
-
-    setLevel(0);
-    setCurrent((prev) => prev + 1);
+    await submitAnswer();
   }
 
   if (loading) {
@@ -77,63 +34,67 @@ export default function Practice() {
   if (branchTasks.length === 0) {
     return (
       <div className="flex min-h-screen items-center justify-center">
-        <div className="text-3xl">No tasks for this branch.</div>
+        <div className="text-3xl underline decoration-[#9AE9DF] decoration-2 underline-offset-4">
+          No tasks for this branch.
+        </div>
       </div>
     );
   }
 
-  if (done) {
-    return (
-      <div className="flex min-h-screen items-center justify-center">
-        <div className="text-[120px]">Done!</div>
-      </div>
-    );
+  if (!currentTask) {
+    return null;
   }
 
   return (
-    <>
-      <div className="max-w-5xl mx-auto mt-8">
-        <form onSubmit={handleSubmit} className="flex flex-col gap-4">
-          <div className="flex flex-col gap-4">
-            <h1 className="text-xl self-start px-3 py-2 bg-[#E8EDF6] border font-[400]">
-              {currentTask.title}
-            </h1>
-            <p className="px-2 font-[400] text-lg whitespace-pre-wrap leading-[1.6]">
-              {currentTask.explanation}
-            </p>
-          </div>
+    <div className="max-w-5xl mx-auto mt-8 pb-24">
+      <form onSubmit={handleSubmit} className="flex flex-col gap-4">
+        <div className="flex flex-col gap-4">
+          <h1 className="px-2 text-3xl underline decoration-[#9AE9DF] decoration-2 underline-offset-4">
+            {currentTask.title}
+          </h1>
+          <p className="px-2 py-4 text-lg whitespace-pre-wrap leading-[1.8]">
+            {currentTask.explanation}
+          </p>
+        </div>
 
-          <div className="flex flex-col">
-            <pre className="bg-[#E8EDF6] p-4 text-lg overflow-x-auto border border-b-0">
-              <code>{currentPrompt}</code>
-            </pre>
-            <textarea
-              className={`h-100 w-full border ${error ? "border-2 border-red-500" : ""} p-4 text-lg font-[monospace]`}
-              value={answer}
-              onChange={(event) => {
-                setError("");
-                setAnswer(event.target.value);
-              }}
-              onKeyDown={(event) => {
-                if (event.key === "Enter" && !event.shiftKey) {
-                  event.preventDefault();
-                  event.currentTarget.form?.requestSubmit();
-                }
-              }}
-            ></textarea>
-          </div>
+        <div ref={editorRef} className="flex flex-col scroll-mt-4">
+          <pre className="bg-[#F5F5F5] p-4 overflow-x-auto border rounded-t border-white border-b-0 leading-[1.6] relative">
+            <div className="absolute top-4 right-4">
+              {totalLevels === 0 ? "No levels" : `${level + 1}/${totalLevels}`}
+            </div>
+            <code>{currentPrompt}</code>
+          </pre>
 
-          <div className="flex flex-row justify-end">
-            <button
-              type="submit"
-              className={`text-xl border px-5 py-2 ${error ? "opacity-50" : ""} bg-[#B8E7FF]`}
-              disabled={error !== "" || loading}
-            >
-              Submit
-            </button>
-          </div>
-        </form>
-      </div>
-    </>
+          <textarea
+            className={`bg-white h-100 w-full border-2 ${error ? "border-[#D85F6F] focus:border-[#D85F6F]" : "border-white focus:border-[#9AE9DF]"} p-4 rounded-b leading-[1.6]`}
+            value={answer}
+            onChange={(event) => {
+              updateAnswer(event.target.value);
+            }}
+            onKeyDown={(event) => {
+              if (event.key === "Enter" && !event.shiftKey) {
+                event.preventDefault();
+                event.currentTarget.form?.requestSubmit();
+              }
+            }}
+            onFocus={() => {
+              editorRef.current?.scrollIntoView({
+                block: "start",
+              });
+            }}
+          />
+        </div>
+
+        <div className="flex flex-row justify-end">
+          <button
+            type="submit"
+            className={`text-xl px-5 py-2 rounded ${error ? "opacity-50" : ""} bg-[#9AE9DF]`}
+            disabled={error !== null || loading || updating}
+          >
+            {updating ? "Saving..." : "Submit"}
+          </button>
+        </div>
+      </form>
+    </div>
   );
 }