+++ server/src/features/tasks/tasks.routes.ts
@@ -1,10 +1,9 @@
 import { Router } from "express";
-import { updateData, getDueTasks } from "./tasks.controller";
+import { getDueTasks, updateData } from "./tasks.controller.js";
 
 const router = Router();
 
 router.get("/", getDueTasks);
-
 router.post("/", updateData);
 
 export default router;